Add lazy load todoActiveList dropdown
